+++ lib/src/controls/better_player_material_controls.dart
@@ -461,7 +461,8 @@ class _BetterPlayerMaterialControlsState
     return StreamBuilder<int>(
       stream: _betterPlayerController.nextVideoTimeStreamController.stream,
       builder: (context, snapshot) {
-        if (snapshot.data != null) {
+        final time = snapshot.data;
+        if (time != null && time > 0) {
           return BetterPlayerMaterialClickableWidget(
             onTap: () {
               _betterPlayerController.playNextVideo();
@@ -477,7 +478,7 @@ class _BetterPlayerMaterialControlsState
                 child: Padding(
                   padding: const EdgeInsets.all(12),
                   child: Text(
-                    "${_betterPlayerController.translations.controlsNextVideoIn} ${snapshot.data} ...",
+                    "${_betterPlayerController.translations.controlsNextVideoIn} $time ...",
                     style: const TextStyle(color: Colors.white),
                   ),
                 ),