+++ lib/src/core/better_player.dart
@@ -117,12 +117,6 @@ class _BetterPlayerState extends State<BetterPlayer>
 
     WidgetsBinding.instance.removeObserver(this);
     widget.controller.removeListener(onFullScreenChanged);
-
-    ///Controller from list widget must be dismissed manually
-    if (widget.controller.betterPlayerPlaylistConfiguration == null) {
-      widget.controller.dispose();
-    }
-
     super.dispose();
   }