+++ lib/src/list/better_player_list_video_player.dart
@@ -65,7 +65,8 @@ class _BetterPlayerListVideoPlayerState
         playerVisibilityChangedBehavior: onVisibilityChanged,
       ),
       betterPlayerDataSource: widget.dataSource,
-      betterPlayerPlaylistConfiguration: BetterPlayerPlaylistConfiguration(),
+      betterPlayerPlaylistConfiguration:
+          const BetterPlayerPlaylistConfiguration(),
     );
 
     if (widget.betterPlayerListVideoPlayerController != null) {