+++ lib/src/playlist/better_player_playlist_configuration.dart
@@ -6,9 +6,9 @@ class BetterPlayerPlaylistConfiguration {
   ///Should videos be looped
   final bool loopVideos;
 
-  BetterPlayerPlaylistConfiguration(
-      {this.nextVideoDelay = const Duration(milliseconds: 3000),
-      this.loopVideos = true})
-      : assert(nextVideoDelay != null && nextVideoDelay.inSeconds >= 3,
-            "NextVideoDelay should be at least 3 seconds");
+  const BetterPlayerPlaylistConfiguration({
+    this.nextVideoDelay = const Duration(milliseconds: 3000),
+    this.loopVideos = true,
+  })  : assert(nextVideoDelay != null),
+        assert(loopVideos != null);
 }